// frontend/src/pages/Authenticated/blogrelated/BlogCreator/types.ts
export type BlogType =
  | 'journey'
  | 'successful_experience'
  | 'milestone'
  | 'report_insight'
  | 'failed_initiation'
  | 'consumption'
  | 'answering_question';

export type ContentType = 'micro' | 'short_essay' | 'article';

export type ReportType =
  | 'activity_reports.DailyVillageActivityReport'
  | 'activity_reports.WeeklyVillageActivityReport'
  | 'activity_reports.MonthlyVillageActivityReport'
  | 'activity_reports.DailySubdistrictActivityReport'
  | 'activity_reports.WeeklySubdistrictActivityReport'
  | 'activity_reports.MonthlySubdistrictActivityReport'
  | 'activity_reports.DailyDistrictActivityReport'
  | 'activity_reports.WeeklyDistrictActivityReport'
  | 'activity_reports.MonthlyDistrictActivityReport'
  | 'activity_reports.DailyStateActivityReport'
  | 'activity_reports.WeeklyStateActivityReport'
  | 'activity_reports.MonthlyStateActivityReport'
  | 'activity_reports.DailyCountryActivityReport'
  | 'activity_reports.WeeklyCountryActivityReport'
  | 'activity_reports.MonthlyCountryActivityReport'
  | 'report.VillageDailyReport'
  | 'report.VillageWeeklyReport'
  | 'report.VillageMonthlyReport'
  | 'report.SubdistrictDailyReport'
  | 'report.SubdistrictWeeklyReport'
  | 'report.SubdistrictMonthlyReport'
  | 'report.DistrictDailyReport'
  | 'report.DistrictWeeklyReport'
  | 'report.DistrictMonthlyReport'
  | 'report.StateDailyReport'
  | 'report.StateWeeklyReport'
  | 'report.StateMonthlyReport'
  | 'report.CountryDailyReport'
  | 'report.CountryWeeklyReport'
  | 'report.CountryMonthlyReport'
  | 'report.CumulativeReport'
  | 'report.OverallReport';